Modificación para que el parser del ej1 tome long long int

+++ tp1/ej1/include/parser.h
@@ -5,6 +5,6 @@
 
 using namespace std;
 
-list<int> parsearInstancias(string entrada);
+list<long long int> parsearInstancias(string entrada);
 
 #endif //__PARSER__
